# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth",
        "description": "Document about Authentification ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/podcasts/create": {
            "post": {
                "tags": [
                    "create podcasts"
                ],
                "summary": "Add podcasts",
                "operationId": "f60ae6bccb4f17d7469bf6e52d6df09b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "catégorie",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "dev"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "text tets"
                                    },
                                    "catégorie": {
                                        "type": "string",
                                        "example": "devlo"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https//images/logo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "podcast created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcasts created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "podcast not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcast not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/podcasts/update/{id}": {
            "put": {
                "tags": [
                    "update podcasts"
                ],
                "summary": "update podcasts",
                "operationId": "b8fdc2490ce1ba2c9a0bf4b989da61b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the podcast to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "catégorie",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "dev"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "text tets"
                                    },
                                    "catégorie": {
                                        "type": "string",
                                        "example": "devlo"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https//images/logo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "podcast update successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcasts update successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "podcast not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcast not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/podcasts/delete/{id}": {
            "delete": {
                "tags": [
                    "delete podcast"
                ],
                "summary": "delete podcast",
                "operationId": "0387bb614be3ec647e72c2d5a69e0793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the podcast to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "podcast delete successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcasts delete successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "podcast not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "podcasts not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}